restructure further
