--- packages/cli/CHANGELOG.md
@@ -1,29 +1,0 @@
-# @anywaydata/cli
-
-## 2.1.0
-
-### Minor Changes
-
-- added support for generating data driven test code scaffolding
-
-### Patch Changes
-
-- Updated dependencies
-  - @anywaydata/core@1.1.0
-
-## 2.0.0
-
-### Major Changes
-
-- Changed API structure to be versioned and additional security controls
-
-### Patch Changes
-
-- Updated dependencies
-  - @anywaydata/core@1.0.0
-
-## 1.0.0
-
-### Major Changes
-
-- First release to npm
